Conflicts:
	Source/UI/SDSearchDisplayController.m

+++ Source/Foundation/UIDevice+machine.m
@@ -85,7 +85,7 @@
         }
     }
 
-    return version;
+    return (uint32_t)version;
 }
 
 + (NSInteger)systemMajorVersion
